+++ datumbox-framework-core/src/test/java/com/datumbox/framework/core/machinelearning/ensemblelearning/BootstrapAggregatingTest.java
@@ -23,7 +23,7 @@ import com.datumbox.framework.core.machinelearning.classification.MultinomialNai
 import com.datumbox.framework.core.machinelearning.modelselection.metrics.ClassificationMetrics;
 import com.datumbox.framework.core.machinelearning.modelselection.Validator;
 import com.datumbox.framework.core.machinelearning.modelselection.splitters.KFoldSplitter;
-import com.datumbox.framework.core.machinelearning.preprocessing.CornerConstraintsEncoder;
+import com.datumbox.framework.core.machinelearning.preprocessing.OneHotEncoder;
 import com.datumbox.framework.core.machinelearning.preprocessing.MinMaxScaler;
 import com.datumbox.framework.tests.Constants;
 import com.datumbox.framework.tests.Datasets;
@@ -66,8 +66,8 @@ public class BootstrapAggregatingTest extends AbstractTest {
         numericalScaler.fit_transform(trainingData);
         numericalScaler.save(storageName);
 
-        CornerConstraintsEncoder.TrainingParameters ceParams = new CornerConstraintsEncoder.TrainingParameters();
-        CornerConstraintsEncoder categoricalEncoder = MLBuilder.create(ceParams, configuration);
+        OneHotEncoder.TrainingParameters ceParams = new OneHotEncoder.TrainingParameters();
+        OneHotEncoder categoricalEncoder = MLBuilder.create(ceParams, configuration);
 
         categoricalEncoder.fit_transform(trainingData);
         categoricalEncoder.save(storageName);
@@ -99,7 +99,7 @@ public class BootstrapAggregatingTest extends AbstractTest {
 
 
         numericalScaler = MLBuilder.load(MinMaxScaler.class, storageName, configuration);
-        categoricalEncoder = MLBuilder.load(CornerConstraintsEncoder.class, storageName, configuration);
+        categoricalEncoder = MLBuilder.load(OneHotEncoder.class, storageName, configuration);
         instance = MLBuilder.load(BootstrapAggregating.class, storageName, configuration);
 
         numericalScaler.transform(validationData);